O(n) es11.js $option = "value" sets SELECT.OPTION
So now $option can be get/set.

Note that $options = [array] is still missing.
So multiselect is NOT supported by now.

Warning!  For now this is O(n) where n is the number of present options.

With observing SELECT and tracing OPTIONs, this could be changed to O(1).
But this seems to be overkill.

Does anybody have that many OPTIONs and selects them so frequently that this eats much CPU?
Also this would eat RAM for tracking.  So either RAM or CPU.  This uses CPU.

Perhaps best would be to implement both:  Some caching object which speeds up $option selection
if this is require, which then eats RAM, which then complements this here.