es11.js PO() {p:promise, ok, ko}
PO() returns a Promise and the ok/ko functions.

Due to async I quite often need something like that:

async function()
{
  let ok,ko;
  p = new Promise((o,k) => { ok=o; ko=k });
  await some_async_processing(p,ok,ko);
  return p;
}

Following is much cleaner for me:

async function wrap(p)	// this wrapper can be generic!
{
  await some_async_processing(p.p, p.ok, p.ko);
  // perhaps return something different here or throw
  return p.p;
}
function()
{
  return wrap(PO());
}

or this can even collapse to something like (async and await
allow here for quickly changing all the real processing steps):

async function()
{
  return await slightly_tweaked_async_processing(PO())
}

The Alternative does not only look ugly and is very redundant to type,
it also is not always applicable if the processing wants to access
the newly crated promise as well.  This makes changing the code very difficult:

function()
{
  return new Promise(async (ok,ko) =>
    {
      // how to access p here?
      await some_async_processing(ok,ko);
    })
}