Semaphore() improved
The first arg can be a function.  It must return the current effective max number to run parallel.
The primary purpose for this is to implement something like a progress bar.
The secondary purpose is to have some easy stop/continue, as if the function returns a Promise,
it is awaited until more can run.

Also some additional function on the returned Semaphore() allow to monitor and control it.